Revert "Replace build id in Pages data routes with deployment id (#88959)"
This reverts commit 6766a270c410459909a791ea4e2585d751cf0742.
