+++ packages/next/src/client/page-loader.ts
@@ -18,7 +18,6 @@ import {
   DEV_CLIENT_MIDDLEWARE_MANIFEST,
 } from '../shared/lib/constants'
 import { resolvePromiseWithTimeout } from './lib/promise'
-import { getDeploymentId } from '../shared/lib/deployment-id'
 
 declare global {
   interface Window {
@@ -171,12 +170,8 @@ export default class PageLoader {
         removeTrailingSlash(addLocale(path, locale)),
         '.json'
       )
-
-      // We check the response header instead
       return addBasePath(
-        getDeploymentId()
-          ? `/_next/data${dataRoute}${search}`
-          : `/_next/data/${this.buildId}${dataRoute}${search}`,
+        `/_next/data/${this.buildId}${dataRoute}${search}`,
         true
       )
     }
