+++ packages/next/src/lib/load-custom-routes.ts
@@ -6,7 +6,6 @@ import { escapeStringRegexp } from '../shared/lib/escape-regexp'
 import { tryToParsePath } from './try-to-parse-path'
 import { allowedStatusCodes } from './redirect-status'
 import { isFullStringUrl } from './url'
-import { NEXT_NAV_DEPLOYMENT_ID_HEADER } from './constants'
 
 export type RouteHas =
   | {
@@ -747,25 +746,13 @@ export default async function loadCustomRoutes(
     )
   }
 
-  if (config.deploymentId) {
-    if (config.experimental?.useSkewCookie) {
-      headers.unshift({
-        source: '/:path*',
-        headers: [
-          {
-            key: 'Set-Cookie',
-            value: `__vdpl=${config.deploymentId}; Path=/; HttpOnly`,
-          },
-        ],
-      })
-    }
-
+  if (config.experimental?.useSkewCookie && config.deploymentId) {
     headers.unshift({
-      source: '/_next/data/(.*)',
+      source: '/:path*',
       headers: [
         {
-          key: NEXT_NAV_DEPLOYMENT_ID_HEADER,
-          value: config.deploymentId,
+          key: 'Set-Cookie',
+          value: `__vdpl=${config.deploymentId}; Path=/; HttpOnly`,
         },
       ],
     })
