+++ packages/next/src/server/lib/router-utils/build-data-route.test.ts
@@ -1,59 +1,29 @@
 import { buildDataRoute } from './build-data-route'
 
 describe('buildDataRoute', () => {
-  describe('with build id', () => {
-    it('should build a dynamic data route', () => {
-      const dataRoute = buildDataRoute('/[...slug]', '123', '')
-      expect(dataRoute).toMatchInlineSnapshot(`
-       {
-         "dataRouteRegex": "^/_next/data/123/(.+?)\\.json$",
-         "namedDataRouteRegex": "^/_next/data/123/(?<nxtPslug>.+?)\\.json$",
-         "page": "/[...slug]",
-         "routeKeys": {
-           "nxtPslug": "nxtPslug",
-         },
-       }
-      `)
-    })
-
-    it('should build a static data route', () => {
-      const dataRoute = buildDataRoute('/about', '123', '')
-      expect(dataRoute).toMatchInlineSnapshot(`
-       {
-         "dataRouteRegex": "^/_next/data/123/about\\.json$",
-         "namedDataRouteRegex": undefined,
-         "page": "/about",
-         "routeKeys": undefined,
-       }
-      `)
-    })
+  it('should build a dynamic data route', () => {
+    const dataRoute = buildDataRoute('/[...slug]', '123')
+    expect(dataRoute).toMatchInlineSnapshot(`
+     {
+       "dataRouteRegex": "^/_next/data/123/(.+?)\\.json$",
+       "namedDataRouteRegex": "^/_next/data/123/(?<nxtPslug>.+?)\\.json$",
+       "page": "/[...slug]",
+       "routeKeys": {
+         "nxtPslug": "nxtPslug",
+       },
+     }
+    `)
   })
 
-  describe('with deployment id', () => {
-    it('should build a dynamic data route', () => {
-      const dataRoute = buildDataRoute('/[...slug]', '123', 'ABCDEF')
-      expect(dataRoute).toMatchInlineSnapshot(`
-       {
-         "dataRouteRegex": "^/_next/data/(.+?)\\.json$",
-         "namedDataRouteRegex": "^/_next/data/(?<nxtPslug>.+?)\\.json$",
-         "page": "/[...slug]",
-         "routeKeys": {
-           "nxtPslug": "nxtPslug",
-         },
-       }
-      `)
-    })
-
-    it('should build a static data route', () => {
-      const dataRoute = buildDataRoute('/about', '123', 'ABCDEF')
-      expect(dataRoute).toMatchInlineSnapshot(`
-       {
-         "dataRouteRegex": "^/_next/data/about\\.json$",
-         "namedDataRouteRegex": undefined,
-         "page": "/about",
-         "routeKeys": undefined,
-       }
-      `)
-    })
+  it('should build a static data route', () => {
+    const dataRoute = buildDataRoute('/about', '123')
+    expect(dataRoute).toMatchInlineSnapshot(`
+     {
+       "dataRouteRegex": "^/_next/data/123/about\\.json$",
+       "namedDataRouteRegex": undefined,
+       "page": "/about",
+       "routeKeys": undefined,
+     }
+    `)
   })
 })
