+++ packages/next/src/server/lib/router-utils/setup-dev-bundler.ts
@@ -1088,11 +1088,7 @@ async function startWatcher(
         const dataRoutes: typeof opts.fsChecker.dynamicRoutes = []
 
         for (const page of sortedRoutes) {
-          const route = buildDataRoute(
-            page,
-            'development',
-            nextConfig.deploymentId
-          )
+          const route = buildDataRoute(page, 'development')
           const routeRegex = getNamedRouteRegex(route.page, {
             prefixRouteKeys: true,
           })
