+++ test/production/required-server-files-ssr-404/test/index.test.ts
@@ -28,6 +28,7 @@ describe('Required Server Files', () => {
       let server: Awaited<ReturnType<typeof startApp>>
       let appPort: number
 
+      let buildId: string
       let requiredFilesManifest: any
 
       // This suite unfortunately requires a very bespoke setup to set NEXT_PRIVATE_TEST_HEADERS and
@@ -35,6 +36,8 @@ describe('Required Server Files', () => {
       beforeAll(async () => {
         await next.build()
         await next.deleteFile('pages')
+
+        buildId = (await next.readFile(join('.next/BUILD_ID'))).trim()
 
         requiredFilesManifest = await next.readJSON(
           join('.next/required-server-files.json')
@@ -193,7 +196,7 @@ describe('Required Server Files', () => {
         const { pageProps: data4 } = JSON.parse(
           await renderViaHTTP(
             appPort,
-            `/_next/data/${next.buildId}/fallback/third.json`,
+            `/_next/data/${buildId}/fallback/third.json`,
             undefined,
             withInvocationId()
           )
@@ -316,7 +319,7 @@ describe('Required Server Files', () => {
       it('should return data correctly with x-matched-path', async () => {
         const res = await fetchViaHTTP(
           appPort,
-          `/_next/data/${next.buildId}/dynamic/first.json?nxtPslug=first`,
+          `/_next/data/${buildId}/dynamic/first.json?nxtPslug=first`,
           undefined,
           withInvocationId({
             headers: {
@@ -332,11 +335,11 @@ describe('Required Server Files', () => {
 
         const res2 = await fetchViaHTTP(
           appPort,
-          `/_next/data/${next.buildId}/fallback/[slug].json`,
+          `/_next/data/${buildId}/fallback/[slug].json`,
           undefined,
           withInvocationId({
             headers: {
-              'x-matched-path': `/_next/data/${next.buildId}/fallback/[slug].json`,
+              'x-matched-path': `/_next/data/${buildId}/fallback/[slug].json`,
               'x-now-route-matches': 'nxtPslug=second',
             },
           })
@@ -502,7 +505,7 @@ describe('Required Server Files', () => {
       it('should return data correctly with x-matched-path for optional catch-all route', async () => {
         const res = await fetchViaHTTP(
           appPort,
-          `/_next/data/${next.buildId}/catch-all.json`,
+          `/_next/data/${buildId}/catch-all.json`,
           undefined,
           withInvocationId({
             headers: {
@@ -518,11 +521,11 @@ describe('Required Server Files', () => {
 
         const res2 = await fetchViaHTTP(
           appPort,
-          `/_next/data/${next.buildId}/catch-all/[[...rest]].json`,
+          `/_next/data/${buildId}/catch-all/[[...rest]].json`,
           undefined,
           withInvocationId({
             headers: {
-              'x-matched-path': `/_next/data/${next.buildId}/catch-all/[[...rest]].json`,
+              'x-matched-path': `/_next/data/${buildId}/catch-all/[[...rest]].json`,
               'x-now-route-matches': 'nxtPrest=hello&rest=hello',
             },
           })
@@ -535,11 +538,11 @@ describe('Required Server Files', () => {
 
         const res3 = await fetchViaHTTP(
           appPort,
-          `/_next/data/${next.buildId}/catch-all/[[...rest]].json`,
+          `/_next/data/${buildId}/catch-all/[[...rest]].json`,
           undefined,
           withInvocationId({
             headers: {
-              'x-matched-path': `/_next/data/${next.buildId}/catch-all/[[...rest]].json`,
+              'x-matched-path': `/_next/data/${buildId}/catch-all/[[...rest]].json`,
               'x-now-route-matches': 'nxtPrest=hello/world&rest=hello/world',
             },
           })
